added device_service
